refactor: Spark 제거 후 Spring으로 전환
