+++ src/main/java/chess/dao/GameDao.java
@@ -1,100 +1,48 @@
 package chess.dao;
 
-import static chess.dao.Connector.getConnection;
+import chess.domain.state.State;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
 
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
+@Repository
 public class GameDao {
 
-    public int findGameId() {
-        final String sql = "select id from game order by id DESC LIMIT 1";
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            return loadGameIdFromDB(statement);
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
+    public void save(String state) {
+        String sql = "insert into game(state) values (?)";
+        jdbcTemplate.update(sql, state);
     }
 
-    private int loadGameIdFromDB(final PreparedStatement statement) throws SQLException {
-        try (ResultSet resultSet = statement.executeQuery()) {
-            return checkExistId(resultSet);
-        }
+    public State findState() {
+        String sql = "select state from game";
+        return jdbcTemplate.queryForObject(
+                sql,
+                (rs, rowNum) -> {
+                    State state = State.of(rs.getString("state"));
+                    return state;
+                });
     }
 
-    private int checkExistId(ResultSet resultSet) throws SQLException {
-        if (!resultSet.next()) {
-            throw new RuntimeException();
-        }
-        return resultSet.getInt("id");
+    public Long findId() {
+        String sql = "select id from game order by id DESC LIMIT 1";
+        return jdbcTemplate.queryForObject(sql, Long.class);
     }
 
-    public String findGameState() {
-        final String sql = "select * from game order by id DESC LIMIT 1";
-
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            return loadStateFromDB(statement);
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
+    public int update(String state, Long gameId) {
+        String sql = "update game set state=? where id=?";
+        return jdbcTemplate.update(sql, state, gameId);
     }
 
-    private String loadStateFromDB(final PreparedStatement statement) throws SQLException {
-        try (final ResultSet resultSet = statement.executeQuery()) {
-            return checkExistState(resultSet);
-        }
+    public int delete() {
+        String sql = "delete from game";
+        return jdbcTemplate.update(sql);
     }
 
-    private String checkExistState(ResultSet resultSet) throws SQLException {
-        if (!resultSet.next()) {
-            throw new RuntimeException();
-        }
-        return resultSet.getString("state");
-    }
-
-    public void save(final String state) {
-        final String sql = "insert into game (state) values (?)";
-
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            statement.setString(1, state);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    public void update(final String state, int gameId) {
-        final String sql = "update game set state = (?) where id = (?)";
-
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            statement.setString(1, state);
-            statement.setInt(2, gameId);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    public void delete() {
-        final String sql = "delete from game";
-
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    public void delete(int id) {
-        final String sql = "delete from game where id = (?)";
-
-        try (final PreparedStatement statement = getConnection().prepareStatement(sql)) {
-            statement.setInt(1, id);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
+    public Long findGameCount() {
+        String sql = "select count(*) from game";
+        return jdbcTemplate.queryForObject(sql, Long.class);
     }
 }
