+++ src/test/java/chess/dao/GameDaoTest.java
@@ -3,69 +3,82 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import java.sql.Connection;
+import chess.dao.GameDao;
+import chess.domain.state.BlackRunning;
+import chess.domain.state.State;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.ActiveProfiles;
 
-public class GameDaoTest {
+@ActiveProfiles("test")
+@SpringBootTest(webEnvironment = WebEnvironment.NONE)
+class GameDaoTest {
 
+    @Autowired
     private GameDao gameDao;
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
     @BeforeEach
     void setUp() {
-        gameDao = new GameDao();
+        jdbcTemplate.execute("DROP TABLE board IF EXISTS");
+        jdbcTemplate.execute("DROP TABLE game IF EXISTS");
+        jdbcTemplate.execute("create table game\n"
+                + "(\n"
+                + "    id    int auto_increment\n"
+                + "        primary key,\n"
+                + "    state varchar(20) not null\n"
+                + ");");
+        jdbcTemplate.execute("create table board\n"
+                + "(\n"
+                + "    position varchar(5)  not null,\n"
+                + "    symbol   varchar(10) not null,\n"
+                + "    color    varchar(10) not null,\n"
+                + "    game_id  int         not null,\n"
+                + "    primary key (position, game_id),\n"
+                + "    constraint board_game_id_fk\n"
+                + "        foreign key (game_id) references game (id)\n"
+                + "            on update cascade on delete cascade\n"
+                + ");");
+        jdbcTemplate.update("insert into game(state) values (?)", "WhiteRunning");
     }
 
     @Test
-    void connection() {
-        final Connection connection = Connector.getConnection();
-        assertThat(connection).isNotNull();
-    }
-
-    @Test
-    void save() {
+    @DisplayName("게임을 생성할 수 있다.")
+    void saveGame() {
         assertThatCode(() -> {
-            gameDao.save("BlackRunning");
+            gameDao.save("WhiteRunning");
         }).doesNotThrowAnyException();
-
-        gameDao.delete(gameDao.findGameId());
     }
 
     @Test
-    void findAGameId() {
-        gameDao.save("WhiteRunning");
-
-        assertThat(gameDao.findGameId()).isInstanceOf(Integer.class);
-
-        gameDao.delete(gameDao.findGameId());
+    @DisplayName("게임의 상태를 조회할 수 있다.")
+    void findState() {
+        assertThat(gameDao.findState()).isInstanceOf(State.class);
     }
 
     @Test
-    void findAGameState() {
-        gameDao.save("WhiteRunning");
-        String state = gameDao.findGameState();
-
-        assertThat(state).isEqualTo("WhiteRunning");
-
-        gameDao.delete(gameDao.findGameId());
+    @DisplayName("게임의 일련번호를 조회할 수 있다.")
+    void findId() {
+        assertThat(gameDao.findId()).isEqualTo(1L);
     }
 
     @Test
-    void update() {
-        assertThatCode(() -> {
-            gameDao.save("BlackRunning");
-            gameDao.update("BlackRunning", gameDao.findGameId());
-        }).doesNotThrowAnyException();
-
-        gameDao.delete(gameDao.findGameId());
+    @DisplayName("게임의 상태를 업데이트할 수 있다.")
+    void updateGameByState() {
+        gameDao.update("BlackRunning", 1L);
+        assertThat(gameDao.findState()).isInstanceOf(BlackRunning.class);
     }
 
     @Test
-    void delete() {
-        gameDao.save("BlackRunning");
-
-        assertThatCode(() -> {
-            gameDao.delete(gameDao.findGameId());
-        }).doesNotThrowAnyException();
+    @DisplayName("게임을 제거할 수 있다.")
+    void deleteGame() {
+        assertThat(gameDao.delete()).isEqualTo(1L);
     }
 }
